Move stuff to parser mod
