+++ lib/src/lib.rs
@@ -8,21 +8,24 @@ mod utils;
 
 pub mod constants;
 pub mod iter;
+pub mod parser;
 pub mod value;
 
 #[cfg(feature = "serde")]
 pub mod serde;
 
+#[cfg(feature = "serde")]
+#[doc(inline)]
+pub use crate::serde::from_str;
 #[doc(inline)]
 pub use errors::Error;
+#[doc(inline)]
+pub use parser::{parse_reader, parse_slice, parse_str};
 #[doc(inline)]
 pub use value::Value;
 
 use std::collections::HashMap;
 use std::hash::{BuildHasher, Hash};
-
-use nom::types::CompleteStr;
-use nom::{call, exact, named};
 
 /// Has scalar length
 pub trait ScalarLength {
@@ -89,9 +92,6 @@ pub enum MergeBehaviour {
     /// __Unimplemented__
     TakeLast,
 }
-
-/// A HCL document body
-pub type Body<'a> = value::MapValues<'a>;
 
 impl<T> OneOrMany<T> {
     pub fn len(&self) -> usize {
@@ -504,64 +504,6 @@ impl Default for MergeBehaviour {
     }
 }
 
-named!(
-    pub body(CompleteStr) -> Body,
-    exact!(call!(value::map_values))
-);
-
-/// Parse a HCL string into a [`Value`] which is close to an abstract syntax tree of the
-/// HCL string.
-///
-/// You can opt to merge the parsed body after parsing. The behaviour of merging is determined by
-/// the [`MergeBehaviour`] enum.
-pub fn parse_str(input: &str, merge: Option<MergeBehaviour>) -> Result<Body, Error> {
-    let (remaining_input, unmerged) =
-        body(CompleteStr(input)).map_err(|e| Error::from_err_str(&e))?;
-
-    if !remaining_input.is_empty() {
-        Err(Error::Bug(format!(
-            r#"Input was not completely parsed:
-Input: {},
-Remaining: {}
-"#,
-            input, remaining_input
-        )))?
-    }
-
-    let pairs = match merge {
-        None => unmerged,
-        Some(MergeBehaviour::Error) => unmerged.merge()?,
-        Some(_) => unimplemented!("Not implemented yet"),
-    };
-
-    Ok(pairs)
-}
-
-/// Parse a HCL string from a IO stream reader
-///
-/// The entire IO stream has to be buffered in memory first before parsing can occur.
-///
-/// When reading from a source against which short reads are not efficient, such as a
-/// [`File`](std::fs::File), you will want to apply your own buffering because the library
-/// will not buffer the input. See [`std::io::BufReader`].
-pub fn parse_reader<R: std::io::Read>(
-    mut reader: R,
-    merge: Option<MergeBehaviour>,
-) -> Result<Body<'static>, Error> {
-    let mut buffer = String::new();
-    reader.read_to_string(&mut buffer)?;
-
-    // FIXME: Can we do better? We are allocating twice. Once for reading into a buffer
-    // and second time calling `as_owned`.
-    Ok(parse_str(&buffer, merge)?.as_owned())
-}
-
-/// Parse a HCL string from a slice of bytes
-pub fn parse_slice(bytes: &[u8], merge: Option<MergeBehaviour>) -> Result<Body, Error> {
-    let input = std::str::from_utf8(bytes)?;
-    parse_str(input, merge)
-}
-
 #[cfg(test)]
 pub(crate) mod fixtures {
     pub static ALL: &[&str] = &[
@@ -581,25 +523,4 @@ pub(crate) mod fixtures {
     pub static SIMPLE_MAP: &str = include_str!("../fixtures/simple_map.hcl");
     pub static SINGLE: &str = include_str!("../fixtures/single.hcl");
     pub static STRINGS: &str = include_str!("../fixtures/strings.hcl");
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn strings_are_parsed_correctly_unmerged() {
-        for string in fixtures::ALL {
-            let parsed = parse_str(string, None).unwrap();
-            assert!(parsed.is_unmerged());
-        }
-    }
-
-    #[test]
-    fn strings_are_parsed_correctly_merged() {
-        for string in fixtures::ALL {
-            let parsed = parse_str(string, Some(MergeBehaviour::Error)).unwrap();
-            assert!(parsed.is_merged());
-        }
-    }
 }
